Implement directory record creation
Also, improve handling of file identifiers, so that rather than
hardcoding an assumption that we'll always be using d-characters, allow
for using d1-characters in future.

Directory record creation consists of walking a filesystem and producing
records for each directory and file. These have to occur in a particular
order as specified by the spec.

For now, we haven't handled locating files or directories on the ISO
disk, and simply leave the location fields as zeros.