Massively refactor codebase into separate packages
This is a gigantic commit, and I absolutely should have shown better Git
hygiene here. However, the codebase was not previously amenable to a
partial refactoring, which is a testament as to why this needed to
happen!

In the previous iteration, everything was inside a single package. From
a testability perspective, this is a bit of a nightmare: it's not
necessarily clear what the given contract of certain parts of the
codebase should be, and hence unit tests were at best spotty and at
worst missed key requirements. From a maintainability perspective, this
was also bad, as there were lots of implementation-specific dependencies
everywhere -- and refactoring even a small part of the codebase could
break the others.

This refactor splits the code into packages with distinct
responsibilities, and makes the contract of these packages clear with
exported symbols (instead of a litter of unexported symbols being used
everywhere). These packages are:

* spec -- holds types and functions representing the ECMA-119
  specification. This allows for spec implementations to focus solely on
  the implementation side of things, primarily by abstracting most
  byte-based/encoding/marshaling operations and providing concrete types
  and interfaces to return and implement.

  This has the added benefit of making it much easier to create
  alternative implementations in the future, and to give implementations
  a common language with which to swap parts in and out.

  Unit tests for this package are not yet implemented, but unit tests
  for the builder package _are_, and are much more focused in their
  aims: rather than having to check both that the builder functionally
  does what it's supposed to (for instance, produces path table records
  in the correct order) **and** produces correct binary data, they can
  now simply test the former. Similarly, when we write the spec unit
  tests, they can simply test the latter, and not worry about the
  former.

* encode -- logic for converting Go types into spec primitive types.
  For example, this package contains logic to encode file identifiers,
  datetimes, numbers, etc.

  Again, the primary benefit here is that the contract of this package
  is much narrower in scope and easier to test, and thus more
  maintainable as it's far less likely that we'll miss key requirements
  in the acceptance tests.

  For example, rather than testing the final binary encoding of
  functions in this package, we can check that the spec types they
  produce are what we expect. Then, the unit tests in the spec package
  can ensure the conversion from spec type to binary is correct.

* builder -- this is the meat of the operation. The builder is more
  loosely-defined, but broadly aims to allow the user to build a full
  ISO by providing the relevant building blocks. Most types in this
  package will have some means of producing a spec type, which have
  well-defined ways of being written to some io.Writer. Thus, the user
  can produce an ISO image!

  The separation of the builder from the spec allows us to separate our
  expectations for what each should do, and hence we can be assured that
  if we break some test in the builder package, we're primarily breaking
  our implementation of the spec -- and that the fault lies either in
  our assumptions about what that implementation should be, or in the
  implementation itself. Notably, we can be confident that we haven't
  broken the spec itself, or how we encode data for the spec. In the
  iteration before this commit, we could break a test and have no idea
  whether, for instance, our logic for ordering path table records was
  incorrect, or our encoding of path table records was incorrect.

Of course, the best laid plans of mice and men often go awry, and hence
these packages may change and evolve over time. However, I think this is
a vast improvement over the previous iteration.

Additionally, this process has revealed some flaws in our logic, which
came about solely through thinking about the contracts of each package
and what the unit tests should actually be testing. I think this speaks
for the utility of the refactor, as we're already finding bugs that
we wouldn't have otherwise!